Modifiche per autorizzazioni ad acesso pagina e azioni

+++ app/Http/Controllers/deletionsController.php
@@ -11,32 +11,41 @@ class deletionsController extends Controller
     //------ELEMENTS_ACTIONS-------//
 
     public function delete_element(Request $request){
-        $id_element = $request->element_idModal;
-        $arrayPhoto = App\PhotoElement::where('element_id', $id_element)->get();
-        $ell = Element::where('id', $id_element)->first();
+        if ( Auth::user()!= null && Auth::user()->group == "Administrator" ) {
+            $id_element = $request->element_idModal;
+            $arrayPhoto = App\PhotoElement::where('element_id', $id_element)->get();
+            $ell = Element::where('id', $id_element)->first();
 
-        PhotoElement::deleteAll($id_element);
-        Element::deleteAll($ell);
-       
+            PhotoElement::deleteAll($id_element);
+            Element::deleteAll($ell);
+        
 
-        $path = $request-> ref;
-        $path = substr($path, 1, strlen($path));
-        return redirect($path);
+            $path = $request-> ref;
+            $path = substr($path, 1, strlen($path));
+            return redirect($path);
+        } else {
+            return abort(403, 'Azione non autorizzata!');
+        }
     }
 
 
     public function delete_element_subcategory(Request $request){
-        $id_element = $request->element_idModal;
-        $arrayPhoto = App\PhotoShowroom::where('element_id', $id_element)->get();
-        $ell = ElementsShowRoom::where('id', $id_element)->first();
-
-        PhotoElement::deleteAll($id_element);
-        Element::deleteAll($ell);
-       
-
-        $path = $request-> ref;
-        $path = substr($path, 1, strlen($path));
-        return redirect($path);
+        if ( Auth::user()!= null && Auth::user()->group == "Administrator" ){
+            $id_element = $request->element_idModal;
+            $arrayPhoto = App\PhotoShowroom::where('element_id', $id_element)->get();
+            $ell = ElementsShowRoom::where('id', $id_element)->first();
+    
+            PhotoElement::deleteAll($id_element);
+            Element::deleteAll($ell);
+           
+    
+            $path = $request-> ref;
+            $path = substr($path, 1, strlen($path));
+            return redirect($path);
+    
+        } else {
+            return abort(403, 'Azione non autorizzata!');
+        }
     }
 
 
@@ -56,9 +65,13 @@ class deletionsController extends Controller
 //------ELEMENT-ACTIONS-END--------//
 
     public function delete_user(Request $request){
-        App\User::where('email', $request->email)->delete();
-        return view('test');
+        if ( Auth::user()!= null && Auth::user()->group == "Administrator" ){
+            App\User::where('email', $request->email)->delete();
+            return view('test');
+        } else {
+            return abort(403, 'Azione non autorizzata!');
         }
+    }
 
     public function delete_news(Request $request){
         App\News::where('name', $request->news)->delete();
@@ -66,24 +79,34 @@ class deletionsController extends Controller
     }
 
     public function delete_category(Request $request){
-        $subcat=App\Subcategory::where('category', $request->category)->get();
+        if ( Auth::user()!= null && Auth::user()->group == "Administrator" ){
+            $subcat=App\Subcategory::where('category', $request->category)->get();
 
-        if (  $subcat != '[]' )
-        return redirect(request()->headers->get('referer').'?openAlert=La%20categoria%20'.$request->category.'%20contiene%20delle%20sottocategorie!');
-        else{
-            App\Category::where('name', $request->category)->delete();
+            if (  $subcat != '[]' )
+            return redirect(request()->headers->get('referer').'?openAlert=La%20categoria%20'.$request->category.'%20contiene%20delle%20sottocategorie!');
+            else{
+                App\Category::where('name', $request->category)->delete();
+            }
+            return redirect(request()->headers->get('referer'));
+
+        } else {
+            return abort(403, 'Azione non autorizzata!');
         }
-        return redirect(request()->headers->get('referer'));
     }
 
     public function delete_subcategory(Request $request){
+        if ( Auth::user()!= null && Auth::user()->group == "Administrator" ){
+            $elements=App\Element::where('subcategories', $request->subcategory)->get();
+            if (  $elements != '[]' )
+                return redirect(request()->headers->get('referer').'?openAlert=La%20sottocategoria%20'.$request->subcategory.'%20contiene%20degli%20elementi!');
+            else{ 
+                App\Subcategory::where('name', $request->subcategory)->delete();
+            }
+            return redirect(request()->headers->get('referer'));
 
-        $elements=App\Element::where('subcategories', $request->subcategory)->get();
-        if (  $elements != '[]' )
-            return redirect(request()->headers->get('referer').'?openAlert=La%20sottocategoria%20'.$request->subcategory.'%20contiene%20degli%20elementi!');
-        else{ 
-            App\Subcategory::where('name', $request->subcategory)->delete();
+        } else {
+            return abort(403, 'Azione non autorizzata!');
         }
-        return redirect(request()->headers->get('referer'));
+
     }
 }
